requirements.txt
atlassian-python-api
pyyaml
requests
